requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
django==5.2.5
django-autoslug==1.9.9
django-cities-light==3.10.1
django-extensions==4.1
django-htmx==1.23.2
dnspython==2.7.0
dotenv==0.9.9
eventlet==0.40.2
greenlet==3.2.4
idna==3.10
kombu==5.5.4
packaging==25.0
progressbar2==4.5.0
prompt-toolkit==3.0.51
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-utils==3.9.1
pytz==2025.2
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.14.1
tzdata==2025.2
unidecode==1.4.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
